+++ extensions/lol-esports/src/lib/hooks/use-schedule.tsx
@@ -12,7 +12,7 @@ export function useSchedule(leagueId: string) {
   };
 }
 
-interface ScheduleResponse {
+export interface ScheduleResponse {
   data: {
     schedule: {
       pages: Pages;